Step 4 of 10
Layout: https://jekyllrb.com/docs/step-by-step/04-layouts/

- Layouts are templates that wrap around your content. They live in a
directory called _layouts.
- This is almost identical to index.html except there’s no front matter
and the content of the page is replaced with a content variable. content
is a special variable which has the value of the rendered content of the
page it’s called on.
- To have other pages use this layout, you can set a layout variable in
front matter. The layout wraps around the content of the page so all you
need in index.html is a variable called layout: <name>.